move view wrapper utils
