+++ external/jwOp.hpp
@@ -3,7 +3,7 @@
 #include <cstdint>
 #include <string>
 
-#include "utils.hpp"
+#include "Quiccir-c/Utils.hpp"
 
 // Mock operation backend (dense batched matmul)
 template <class T>
