+++ include/Quiccir-c/Utils.hpp
